# Improved performance of writing to CSV (20-25%)