--- a/benches/comparison_kernels.rs
+++ b/benches/comparison_kernels.rs
@@ -43,8 +43,8 @@ fn add_benchmark(c: &mut Criterion) {
             b.iter(|| bench_op_scalar(&arr_a, &BooleanScalar::from(Some(true)), Operator::Eq))
         });
 
-        let arr_a = create_string_array::<i32>(size, 0.1, 42);
-        let arr_b = create_string_array::<i32>(size, 0.1, 43);
+        let arr_a = create_string_array::<i32>(size, 4, 0.1, 42);
+        let arr_b = create_string_array::<i32>(size, 4, 0.1, 43);
         c.bench_function(&format!("utf8 2^{}", log2_size), |b| {
             b.iter(|| bench_op(&arr_a, &arr_b, Operator::Eq))
         });