--- a/src/util/bench_util.rs
+++ b/src/util/bench_util.rs
@@ -95,18 +95,23 @@ where
         .collect()
 }
 
-/// Creates an random (but fixed-seeded) array of a given size and null density
-pub fn create_string_array<O: Offset>(size: usize, null_density: f32, seed: u64) -> Utf8Array<O> {
+/// Creates an random (but fixed-seeded) [`Utf8Array`] of a given length, number of characters and null density.
+pub fn create_string_array<O: Offset>(
+    length: usize,
+    size: usize,
+    null_density: f32,
+    seed: u64,
+) -> Utf8Array<O> {
     let mut rng = StdRng::seed_from_u64(seed);
 
-    (0..size)
+    (0..length)
         .map(|_| {
             if rng.gen::<f32>() < null_density {
                 None
             } else {
                 let value = (&mut rng)
                     .sample_iter(&Alphanumeric)
-                    .take(4)
+                    .take(size)
                     .map(char::from)
                     .collect::<String>();
                 Some(value)