--- a/src/util/lexical.rs
+++ b/src/util/lexical.rs
@@ -2,6 +2,15 @@
 #[inline]
 pub fn lexical_to_bytes<N: lexical_core::ToLexical>(n: N) -> Vec<u8> {
     let mut buf = Vec::<u8>::with_capacity(N::FORMATTED_SIZE_DECIMAL);
+    lexical_to_bytes_mut(n, &mut buf);
+    buf
+}
+
+/// Converts numeric type to a `String`
+#[inline]
+pub fn lexical_to_bytes_mut<N: lexical_core::ToLexical>(n: N, buf: &mut Vec<u8>) {
+    buf.clear();
+    buf.reserve(N::FORMATTED_SIZE_DECIMAL);
     unsafe {
         // JUSTIFICATION
         //  Benefit
@@ -13,7 +22,6 @@ pub fn lexical_to_bytes<N: lexical_core::ToLexical>(n: N) -> Vec<u8> {
         let len = lexical_core::write(n, slice).len();
         buf.set_len(len);
     }
-    buf
 }
 
 /// Converts numeric type to a `String`