Currently, we write to CSV by using an `Iterator<Item=Vec<u8>>`. This requires a new allocation per non-null item.

This PR uses [`StreamingIterator`](https://docs.rs/streaming-iterator/0.1.5/streaming_iterator/) to significantly reduce the number of allocations.

A `StreamingIterator` is an Iterator-like trait that allows yielding references of itself, `&[u8]` in this case. We maintain an internal buffer of `Vec<u8>` on the (streaming) iterator and re-use it across items within the array.

In summary, this replaces (1 alloc + write bytes) by a (maybe 1 realloc + write bytes) per non-null item. For types that require a fixed number of bytes (e.g. all our primitive types), it results in a single allocation per array, as opposed to an allocation per non-null item.

```
csv write i32 2^18      time:   [15.828 ms 15.866 ms 15.915 ms]                               
                        change: [-26.127% -25.784% -25.480%] (p = 0.00 < 0.05)
csv write utf8 2^18     time:   [33.100 ms 33.188 ms 33.276 ms]                                
                        change: [-14.104% -13.811% -13.508%] (p = 0.00 < 0.05)
csv write f64 2^18      time:   [23.661 ms 23.702 ms 23.748 ms]                               
                        change: [-20.623% -20.448% -20.239%] (p = 0.00 < 0.05)
```